Consolidate audit CLI helpers
