+++ scinoephile/cli/audit/__init__.py
@@ -3,7 +3,7 @@
 """Command-line interfaces for subtitle audits.
 
 Package hierarchy (modules may import from any above):
-* audit_cli_base / utils
+* audit_cli_base
 * audit_review_cli_base
 * audit_review_cli / audit_review_dual_cli
 / audit_review_trad_cli
